do some renames
