+++ packages/utils/src/clientreport.ts
@@ -1,4 +1,4 @@
-import { ClientReportEnvelope, ClientReportEnvelopeItemHeader, ClientReportEnvelopeItemPayload } from '@sentry/types';
+import { ClientReport, ClientReportEnvelope } from '@sentry/types';
 
 import { createEnvelope } from './envelope';
 import { dateTimestampInSeconds } from './time';
@@ -9,14 +9,14 @@ import { dateTimestampInSeconds } from './time';
  * @param dsn A DSN that can be set on the header. Optional.
  */
 export function createClientReportEnvelope(
-  discarded_events: ClientReportEnvelopeItemPayload['discarded_events'],
+  discarded_events: ClientReport['discarded_events'],
   dsn?: string,
   timestamp?: number,
 ): ClientReportEnvelope {
   const header = dsn ? { dsn } : {};
 
-  const itemHeader: ClientReportEnvelopeItemHeader = { type: 'client_report' };
-  const itemPayload: ClientReportEnvelopeItemPayload = {
+  const itemHeader = { type: 'client_report' };
+  const itemPayload = {
     timestamp: timestamp || dateTimestampInSeconds(),
     discarded_events,
   };
